[ETS] Updating type alias
Issue:
https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/I9U38V

Reason:
Type alias doesn't work correctly.
It should allowed to be declared in function scopes and with the same name as variable that is not an other type alias.

Description:
Added a local scope for type aliases to the function scope to separate them from other variables.
Changes in function scope variable binding

Signed-off-by: Gergo Torok <gergo.torok@huawei.com>