Fix defaultParameter lowering for constructors
Issue:
https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/IBKRH6

Internal issue: #20737

Reason: Invalid CTE was thrown for constructor with default parameter
and a return statement.

Description:
defaultParameter lowering was generated incorrect source code
for constructors: "return this;" instead of "this(undefined);".

Additionally checker went to an infinite recursion if a primary
constructor called a secondary constructor which contains a return.

Test: new runtime tests

Change-Id: Ic2948e37acd32b938f4ea14879d8937458f338d0
Co-authored-by: Martin Sajti <martin.sajti@h-lab.eu>
Signed-off-by: Csaba Osztrogonac <csaba.osztrogonac@h-lab.eu>