Fix invalid overload declaration crash
Issue:
https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/ICMTFJ

Description:
The overload points to an incorrect identifier, and the Tstype of the identifier is Typeerror. When creating the ETSFunctionType for the overload call, these Typeerrors were not filtered, leading to compilation crash.
If the TsType of the identifier in the overload list is not ETSFunctionType, it should be filtered out and not added to the signature.

Tested-by:
ninja tests (passed)
ets_testrunner (passed)

Signed-off-by: zengzengran <zengran2@huawei.com>#