Remove redundant assert
https://gitcode.com/openharmony/arkcompiler_ets_frontend/issues/7692
Internal Issue: 30184

The removed assert is part of a commit (c960cc42ee) that is based on a version of the specification
where the combined use of rest parameters and optional parameters was disallowed.

It is legal though meaningless to declare a function with an optional parameter followed by a rest parameter.

Change-Id: I09d5e50ce8e3ed7d0e01f6b1727cc551a0a42933
Signed-off-by: Robert Sipka <robert.sipka@h-lab.eu>